Fixed export data test
